Tercer Entrega a un 90%

+++ js/appTercerEntrega.js
@@ -24,55 +24,81 @@ ESFUERZO (5MIN) Y QUIERO QUE QUEDE A POSTERIDAD.
 
 //BASE DE CLIENTES
 
-const baseClientes = [
-  {
-    idCliente: 1,
-    nombre: "EDUARDO GONZALEZ",
-    telefono: 61616565,
-    edad: 21,
-    direccion: "avenida siempre viva 120",
-    alergias: "pimienta",
-  },
-  {
-    idCliente: 2,
-    nombre: "jULIO MEDINA",
-    telefono: 61616566,
-    edad: 22,
-    direccion: "avenida siempre viva 121",
-    alergias: "loroco",
-  },
-  {
-    idCliente: 3,
-    nombre: "JOSUE GONZALEZ",
-    telefono: 61616060,
-    edad: 23,
-    direccion: "avenida siempre viva 122",
-  },
-  {
-    idCliente: 4,
-    nombre: "HOMERO SIMPSON",
-    telefono: 70707878,
-    edad: 34,
-    direccion: "avenida siempre viva 123",
-    alergias: "pimienta",
-  },
-  {
-    idCliente: 5,
-    nombre: "GABRIELA GONZALEZ",
-    telefono: 61616000,
-    edad: 27,
-    direccion: "avenida siempre viva 124",
-  },
-  {
-    idCliente: 6,
-    nombre: "ALE",
-    telefono: 78707870,
-    edad: 30,
-    direccion: "avenida siempre viva 125",
-    alergias: "colorante rojo",
-  },
-];
+class baseCliente {
+  constructor(idCliente, telefono, nombre, edad, direccion, alergias) {
+    this.idCliente = idCliente;
+    this.telefono = telefono;
+    this.nombre = nombre;
+    this.edad = edad;
+    this.direccion = direccion;
+    this.alergias = alergias;
+  }
+}
 
+console.log(localStorage.length);
+
+if (localStorage.length < 4) {
+  let baseCliente = [
+    {
+      idCliente: 1,
+      nombre: "EDUARDO GONZALEZ",
+      telefono: 61616565,
+      edad: 21,
+      direccion: "avenida siempre viva 120",
+      alergias: "pimienta",
+    },
+    {
+      idCliente: 2,
+      nombre: "jULIO MEDINA",
+      telefono: 61616566,
+      edad: 22,
+      direccion: "avenida siempre viva 121",
+      alergias: "loroco",
+      alergias: "chile",
+    },
+    {
+      idCliente: 3,
+      nombre: "JOSUE GONZALEZ",
+      telefono: 61616060,
+      edad: 23,
+      direccion: "avenida siempre viva 122",
+      alergias: "colorantes",
+    },
+    {
+      idCliente: 4,
+      nombre: "HOMERO SIMPSON",
+      telefono: 70707878,
+      edad: 34,
+      direccion: "avenida siempre viva 123",
+      alergias: "pimienta",
+    },
+    {
+      idCliente: 5,
+      nombre: "GABRIELA GONZALEZ",
+      telefono: 61616000,
+      edad: 27,
+      direccion: "avenida siempre viva 124",
+      alergias: "chicharon",
+    },
+    {
+      idCliente: 6,
+      nombre: "ALE",
+      telefono: 78707870,
+      edad: 30,
+      direccion: "avenida siempre viva 125",
+      alergias: "colorante rojo",
+    },
+  ];
+  console.log(baseCliente);
+
+  baseClientes = baseCliente;
+} else if (localStorage.length >= 4) {
+  let baseCliente = JSON.parse(localStorage.getItem("baseClientes"));
+  console.log(baseCliente);
+  baseClientes = baseCliente;
+}
+
+console.log(baseClientes);
 class MiCliente {
   constructor(idCliente, telefono, nombre, edad, direccion, alergias) {
     this.idCliente = idCliente;
@@ -84,28 +110,10 @@ class MiCliente {
   }
 }
 
-//Funcion para depurando de valores !logicos
-function recarga() {
-  let entrada = confirm(
-    "Valores invalidos!!!\nTipos de error.\n\nTienes que ingresar cuantas personas (15 personas maximo).\nOrdenaste un numero negativo\nQuieres empezar de nuevo?"
-  );
-
-  if (entrada) {
-    location.reload();
-    throw new Error(
-      "Numero de personas Mayor al esperado, deteniendo el script"
-    );
-  } else {
-    alert("Un gusto atenderte, vuelve pronto");
-    throw new Error(
-      "Numero de personas Mayor al esperado, deteniendo el script"
-    );
-  }
-}
-
 let pupas = [{}, {}, {}, {}, {}];
+
 //FUNCION QUE REVISA SI EXISTE UN PEDIDO ANTERIOR
-if (localStorage.length >= 5) {
+if (localStorage.length >= 2) {
   const pedidoInicial = document.getElementById("datosPedido");
   let revueltas = document.getElementById("rev");
   let queso = document.getElementById("fri");
@@ -113,11 +121,13 @@ if (localStorage.length >= 5) {
   let chicharon = document.getElementById("chich");
   let vegetariana = document.getElementById("vege");
 
-  revueltas.value = Number(localStorage.getItem(`revueltas`));
-  queso.value = Number(localStorage.getItem(`queso`));
-  campesina.value = Number(localStorage.getItem(`campesina`));
-  chicharon.value = Number(localStorage.getItem(`chicharon`));
-  vegetariana.value = Number(localStorage.getItem(`vegetariana`));
+  let pupas = JSON.parse(localStorage.getItem("pupas"));
+  console.log(pupas);
+  revueltas.value = pupas[0];
+  queso.value = pupas[1];
+  campesina.value = pupas[2];
+  chicharon.value = pupas[3];
+  vegetariana.value = pupas[4];
 }
 
 // INGRESANDO VALORES CON EVENTOS.
@@ -164,11 +174,7 @@ botonPupusas.onclick = () => {
   let TotalPrecio = PrecioPupas.reduce((a, b) => a + b, 0);
 
   //ALMACEN DE DATOS PARA NO PERDER EL PEDIDO
-  localStorage.setItem("revueltas", pupas[0]);
-  localStorage.setItem("queso", pupas[1]);
-  localStorage.setItem("campesina", pupas[2]);
-  localStorage.setItem("chicharon", pupas[3]);
-  localStorage.setItem("vegetariana", pupas[4]);
+  localStorage.setItem("pupas", JSON.stringify(pupas));
   localStorage.setItem("TotalPupas", TotalPupas);
   localStorage.setItem("TotalPrecio", TotalPrecio.toFixed(2));
 
@@ -176,50 +182,79 @@ botonPupusas.onclick = () => {
   console.log("TOTAL PUPUSAS = " + TotalPupas);
   console.log("$ " + TotalPrecio.toFixed(2));
 
-  if (
-    pupas[0] == 0 &&
-    pupas[1] == 0 &&
-    pupas[2] == 0 &&
-    pupas[3] == 0 &&
-    pupas[4] == 0
-  ) {
-    //DEJARE SOLO ESTE ALERT PARA PODER ADVERTIR UNA VEZ QUE NO SE INGRESARON DATOS
-    localStorage.clear();
-    alert(
-      "Porfavor, ingresa cuantas pupas y luego pulsa en enviar para poder tomar tu orden"
-    );
-    location.reload();
-  }
-
   let clienteFinal = document.getElementById("nombre");
   let edadCliente = document.getElementById("age");
   let telefonoCleinte = document.getElementById("telefono");
   let direccionCliente = document.getElementById("direccion");
   let alergiaCliente = document.getElementById("alergia");
 
-  console.log("Cliente= "+clienteFinal.value);
-  console.log("Edad= "+edadCliente.value);
-  console.log("Telefono= "+telefonoCleinte.value);
-  console.log("Direccion= "+direccionCliente.value);
-  console.log("Alergias= "+alergiaCliente.value);
+  console.log("Cliente= " + clienteFinal.value);
+  console.log("Edad= " + edadCliente.value);
+  console.log("Telefono= " + telefonoCleinte.value);
+  console.log("Direccion= " + direccionCliente.value);
+  console.log("Alergias= " + alergiaCliente.value);
 
-  clienteNuevo = new MiCliente (baseClientes.length+1,telefonoCleinte.value,clienteFinal.value,edadCliente.value,direccionCliente.value,alergiaCliente.value);
+  clienteNuevo = new MiCliente(
+    baseClientes.length + 1,
+    telefonoCleinte.value,
+    clienteFinal.value,
+    edadCliente.value,
+    direccionCliente.value,
+    alergiaCliente.value
+  );
 
+  if (clienteFinal.value == 0) {
+    alert(
+      "Porfavor, ingresa cuantas pupas y luego tu nombre completo y direccion antes de pulsar el boton\n para enviar tu orden"
+    );
+    location.reload();
+  }
+
+  if (direccionCliente.value == 0) {
+    direccionCliente.value = "Recojer en tienda";
+  }
+
+  if (
+    pupas[0] == 0 &&
+    pupas[1] == 0 &&
+    pupas[2] == 0 &&
+    pupas[3] == 0 &&
+    pupas[4] == 0 &&
+    direccionCliente.value == 0
+  ) {
+    //DEJARE SOLO ESTE ALERT PARA PODER ADVERTIR UNA VEZ QUE NO SE INGRESARON DATOS
+    localStorage.clear();
+    alert(
+      "Porfavor, ingresa cuantas pupas y luego tu direccion antes de pulsar el boton\n para enviar tu orden"
+    );
+    location.reload();
+  }
   botonPupusas.remove();
 
-  if (localStorage.length >= 7) {
+  if (localStorage.length >= 3) {
     const divCliente = document.getElementById("cliente");
     const divContenedor = document.getElementById("datosPedido");
 
     let botonAlmacenarCliente = document.createElement("button");
-    botonAlmacenarCliente.setAttribute("class", "botonDireccion");
-    botonAlmacenarCliente.setAttribute("id", "botonDireccion");
-    botonAlmacenarCliente.textContent = "Almacenar mis datos";
+
+    let buscartelefono = baseClientes.find(
+      (item) => item.telefono === telefonoCleinte.value
+    );
+
+    if (
+      buscartelefono === undefined &&
+      direccionCliente.value != "Recojer en tienda"
+    ) {
+      botonAlmacenarCliente.setAttribute("class", "botonDireccion");
+      botonAlmacenarCliente.setAttribute("id", "botonDireccion");
+      botonAlmacenarCliente.textContent = "Almacenar mis datos";
+    }
     botonAlmacenarCliente.onclick = () => {
       //PUSHEAR LOS DAROS A LA BASE DE DATOS DEL CLIENTES Y ALMACENAR
       baseClientes.push(clienteNuevo);
-
+      localStorage.setItem("baseClientes", JSON.stringify(baseClientes));
       botonAlmacenarCliente.remove();
+      console.log(baseClientes);
     };
     divCliente.appendChild(botonAlmacenarCliente);
 
@@ -252,11 +287,11 @@ botonPupusas.onclick = () => {
       const resumenPupusas = document.createElement("p");
       resumenPupusas.innerText =
         "El resumen de tu pedido es:\n\n" +
-        `${pupas[0]} - Pupusas Revueltas\t - (1.00)\t - $ ${(
-          pupas[0] * 1.0
+        `${pupas[0]} - Pupusas Revueltas\t - (0.90)\t - $ ${(
+          pupas[0] * 0.9
         ).toFixed(2)}\n` +
-        `${pupas[1]} - Pupusas La queso\t - (1.25)\t - $ ${(
-          pupas[1] * 1.25
+        `${pupas[1]} - Pupusas La queso\t - (0.95)\t - $ ${(
+          pupas[1] * 0.95
         ).toFixed(2)}\n` +
         `${pupas[2]} - Pupusas Campesina\t - (1.15)\t - $ ${(
           pupas[2] * 1.15
